FGVC/requirements.txt
matplotlib==3.3.1
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.2
seaborn==0.11.0
split-folders==0.5.1
timm==0.5.4
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.1